requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
appdirs==1.4.4
argcomplete==3.5.3
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
crispy-bootstrap4==2024.1
cryptography==44.0.0
Django==5.0.4
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-push-notifications==3.1.0
djangorestframework==3.15.1
exceptiongroup==1.2.2
Flask==3.1.0
fontawesomefree==6.5.1
frozenlist==1.5.0
gTTS==2.5.4
h11==0.14.0
http-ece==1.2.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.1.0
outcome==1.3.0.post0
phonenumbers==8.13.35
pillow==10.3.0
propcache==0.2.1
psycopg2-binary==2.9.9
py-vapid==1.9.2
pycparser==2.22
PySocks==1.7.1
python-barcode==0.15.1
python-decouple==3.8
python-escpos==3.1
pyttsx3==2.98
pywebpush==2.0.3
PyYAML==6.0.2
qrcode==8.0
requests==2.32.3
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
